pyproject.toml
[project]
name = "cryosight"
version = "0.1.0"
description = "CryoSight knowledge base backend with FastAPI and pgvector"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.0",
    "httpx>=0.28.1",
    "pgvector>=0.3.6",
    "psycopg[binary]>=3.2.0",
    "pydantic-settings>=2.6.1",
    "pypdf>=5.4.0",
    "python-multipart>=0.0.20",
    "sqlalchemy[asyncio]>=2.0.36",
    "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
extend-exclude = ["pipeline"]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
